--- aleph/migrate/versions/aca8cda02c58_create_diagram.py
@@ -1,38 +1,0 @@
-"""Create Diagram table
-
-Revision ID: aca8cda02c58
-Revises: b3ff632002d9
-Create Date: 2019-11-27 07:02:11.988577
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = 'aca8cda02c58'
-down_revision = 'b3ff632002d9'
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('diagram',
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.Column('deleted_at', sa.DateTime(), nullable=True),
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('data', postgresql.JSONB(astext_type=sa.Text()), nullable=True),
-    sa.Column('role_id', sa.Integer(), nullable=True),
-    sa.Column('collection_id', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['collection_id'], ['collection.id'], ),
-    sa.ForeignKeyConstraint(['role_id'], ['role.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_diagram_collection_id'), 'diagram', ['collection_id'], unique=False)
-    op.create_index(op.f('ix_diagram_role_id'), 'diagram', ['role_id'], unique=False)
-
-
-def downgrade():
-    op.drop_index(op.f('ix_diagram_role_id'), table_name='diagram')
-    op.drop_index(op.f('ix_diagram_collection_id'), table_name='diagram')
-    op.drop_table('diagram')
